my test/dashboard-iot-backend-updated/requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
dnspython==2.4.2
filelock==3.12.3
Flask==2.3.3
Flask-Cors==4.0.0
future==0.18.3
gevent==23.9.0.post1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
pycparser==2.21
pymongo==4.5.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
tldextract==3.4.4
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
